requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
elevenlabs==1.13.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.2
pluggy==1.5.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.3
websockets==14.1
